Re #11617. Created a base class for fitting algorithms.
